{"schema":"decodex/commit/1","summary":"split status and activity support modules","authority":"manual"}
